requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
jinja2==3.1.2
python-multipart==0.0.5
tortoise-orm==0.19.2


# dev
pytest==7.2.0
httpx==0.23.1
pytest-asyncio==0.20.3